requirements.txt
Streamlit
streamlit-authenticator
azure-storage-blob
